Conflicts:
	xorg-server/dix/dispatch.c
	xorg-server/hw/xwin/ddraw.h
	xorg-server/hw/xwin/glx/glshim.c
	xorg-server/hw/xwin/winclipboard/xevents.c
	xorg-server/hw/xwin/windialogs.c
	xorg-server/hw/xwin/winmultiwindowshape.c
	xorg-server/hw/xwin/winmultiwindowwindow.c
	xorg-server/hw/xwin/winprefslex.l
	xorg-server/hw/xwin/winshadddnl.c
	xorg-server/hw/xwin/winshadgdi.c
	xorg-server/hw/xwin/winwndproc.c
	xorg-server/mi/miarc.c
	xorg-server/os/connection.c

+++ xorg-server/include/regionstr.h
@@ -125,7 +125,7 @@ RegionEnd(RegionPtr reg)
 }
 
 static inline size_t
-RegionSizeof(int n)
+RegionSizeof(size_t n)
 {
     return (sizeof(RegDataRec) + ((n) * sizeof(BoxRec)));
 }
